% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCtsRNAscore.R
\name{predict_tsRNA}
\alias{predict_tsRNA}
\title{Calculate Gastric Cancer tsRNA Score (GCtsRNAscore)}
\usage{
predict_tsRNA(my_expr)
}
\arguments{
\item{my_expr}{A numeric matrix with rows as samples and columns as genes.}
}
\value{
A data frame with sample names and their tsRNA scores.
}
\description{
This function calculates the GC tsRNA scores based on a given gene expression matrix.
The expression matrix should have rows as samples and columns as genes.
If the matrix is missing required genes, an error will be thrown.
}
